requirements_dev.txt
coverage==4.5.1
flake8==3.5.0
jenkins==1.0.2
mccabe==0.6.1
mock==2.0.0
mysql-connector-python==8.0.11
nose==1.3.7
pbr==4.0.3
Pillow==5.1.0
protobuf==3.5.2.post1
pycodestyle==2.3.1
pyflakes==1.6.0
python-decouple==3.1
six==1.11.0
